requirements.txt
torchaudio
setuptools
pyaml
lightning
matplotlib
seaborn
pandas
einops
tensorboard
mamba-ssm[causal-conv1d]
